move methods in *SearchResultBundle to *SearchDocument
This results in *SearchResultBundle becoming pure data structures instead of classes, and thus exempt-able from testing.
